pyproject.toml
[tool.poetry]
name = "handwriting-ocr"
version = "0.1.0"
description = "A handwriting OCR pipeline with shadow removal preprocessing"
authors = ["sjvrensburg <stefanj@mandela.ac.za>"]
readme = "README.md"
packages = [{include = "handwriting_ocr"}]

[tool.poetry.dependencies]
python = "^3.12.1"
torch = "^2.5.1"
torchvision = "^0.20.1"
torchaudio = "^2.5.1"
transformers = "^4.45.2"
pillow = "^10.2.0"
click = "^8.0.0"
shadow-removal = {git = "https://github.com/sjvrensburg/shadow_removal.git"}
bitsandbytes = "^0.44.1"
accelerate = "^1.0.1"
optimum = "^1.23.2"
safetensors = "^0.4.5"
anthropic = "^0.37.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.0"
isort = "^5.13.0"

[tool.poetry.scripts]
ocr = "handwriting_ocr.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
